+++ input/sdi/x86/sdi.h
@@ -33,7 +33,7 @@ void obe_v210_planar_unpack_unaligned_avx( const uint32_t *src, uint16_t *y, uin
 void obe_v210_planar_unpack_aligned_ssse3( const uint32_t *src, uint16_t *y, uint16_t *u, uint16_t *v, int width );
 void obe_v210_planar_unpack_aligned_avx( const uint32_t *src, uint16_t *y, uint16_t *u, uint16_t *v, int width );
 
-void obe_v210_line_to_nv20_ssse3( uint16_t *dsty, intptr_t i_dsty, uint16_t *dstc, intptr_t i_dstc, uint32_t *src, intptr_t i_src, int width, int h );
-void obe_v210_line_to_nv20_avx( uint16_t *dsty, intptr_t i_dsty, uint16_t *dstc, intptr_t i_dstc, uint32_t *src, intptr_t i_src, int width, int h );
+void obe_v210_line_to_nv20_ssse3( uint16_t *dsty, intptr_t i_dsty, uint16_t *dstc, intptr_t i_dstc, uint32_t *src, intptr_t i_src, intptr_t width, intptr_t h );
+void obe_v210_line_to_nv20_avx( uint16_t *dsty, intptr_t i_dsty, uint16_t *dstc, intptr_t i_dstc, uint32_t *src, intptr_t i_src, intptr_t width, intptr_t h );
 
 #endif
